Created the program, and tried to make a README.md. Hopefully this is good.
Code is in intro_to_python.py. README.md reads somewhat like an essay with headings. Has mentions of requirements.txt (and how to make it) and how to run intro_to_python.py from the command line. NumPy was installed and should be visible in the requirements.txt file.